Don't delete _i18n in beforeDestroy (#880)
Avoid that attempts to call methods on _i18n after destroying a
component fail. This is for example relevant for long running
async methods that were called before the component got destroyed
but terminate after the destruction and then try to call an i18n
method.
Note that _18n doesn't need to be cleared manually. It gets
garbage collected when the destroyed component gets destroyed.

Fixes #879